Unified Day 18 code to use same function part 1 and 2

+++ 2016/Day_24.hpp
@@ -56,14 +56,14 @@ private:
 			{
 				if (pos.y != 2)
 				{
-					int32_t y = pos.y * 2 - 2;
+					int64_t y = pos.y * 2 - 2;
 					for (int x = 0; x < 5; ++x)
 						recursiveNeighbours.push_back(v3(x, y, input.z + 1));
 				}
 
 				if (pos.x != 2)
 				{
-					int32_t x = pos.x * 2 - 2;
+					int64_t x = pos.x * 2 - 2;
 					for (int y = 0; y < 5; ++y)
 						recursiveNeighbours.push_back(v3(x, y, input.z + 1));
 				}
